Expose symbolic indicator discovery
